fix: 存在しないドキュメントにアクセスするとAPIコールが止まらない問題を修正
