+++ src/components/App.tsx
@@ -18,6 +18,7 @@ import GroupNameForm from './pages/group/GroupNameForm';
 import GroupNameFormName from './pages/group/GroupNameFormName';
 import NotFound from './pages/NotFound';
 import { AlertProvider } from '../contexts/AlertContext';
+import DocumentName from './pages/document/DocumentName';
 
 const PrivateRoute: React.FC<RouteProps> = ({ ...props }) => {
   const auth = useAuth();
@@ -52,6 +53,7 @@ const App = (): JSX.Element => (
               <UnAuthRoute exact path="/login" component={Login} />
               <PrivateRoute exact path={Pages.index.href} component={Index} />
               <PrivateRoute exact path={Pages.document.href} component={Document} />
+              <PrivateRoute exact path={Pages.documentName.href} component={DocumentName} />
               <PrivateRoute exact path={Pages.group.href} component={Group} />
               <PrivateRoute exact path={Pages.groupName.href(':groupName')} component={GroupName} />
               <PrivateRoute exact path={Pages.groupNameForm.href(':groupName')} component={GroupNameForm} />
